requirements.txt
Babel==1.3

Factory==1.2
Markdown==2.0.3
MySQL-Connector-Python==1.1.6
PyYAML==3.11
RapidSMS==0.17.0
South==0.8.4

Whoosh==2.6.0
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.2
billiard==2.7.3.34
bottle==0.11.6
celery==3.0.24
cmsplugin-filer==0.9.5
coverage==3.7.1
coveralls==0.4.2
distribute==0.6.13
dj-database-url==0.3.0
django-admin-sortable==1.7
django-appconf==0.6
django-blog-zinnia==0.14.1
django-bootstrap-pagination==1.5.0
django-celery==3.0.23
django-classy-tags==0.5.1
django-cms==3.0
django-compressor==1.3
django-contrib-comments==1.5
django-debug-toolbar==1.0.1
django-debug-toolbar-template-timings==0.6.1
django-extensions==1.3.3
django-extra-views==0.6.4
django-filer==0.9.5
django-haystack==2.1.0
django-i18nurls==0.7
django-localflavor==1.0
django-mptt==0.5.5
django-multilingual-model==0.6
django-nose==1.2

django-pagination==1.0.7
django-polymorphic==0.5.4
-e git+https://github.com/dokterbob/django-portfolio.git@9a7acf3daa58d6543fa25a5dd03e2eaac64f57c6#egg=django_portfolio-master
django-redis==3.6.2
django-reversion==1.8.0
django-robokassa==1.2
django-rq==0.6.2
django-sekizai==0.7
django-selectable==0.7.0
django-smsaero==0.9.5
django-staticfiles-isotope==1.5.18
django-tables==0.2
django-tables2==0.14.0
django-tagging==0.3.2
django-treebeard==2.0
django-widget-tweaks==1.3
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.2
djangocms-installer==0.5.2
djappsettings==0.1.0
djtables==0.1.2
docopt==0.6.1

feedparser==5.1.3
html5lib==0.999
kombu==2.5.16
mercurial==2.9
mock==1.0.1
nose==1.3.3
pbs==0.110
phonenumbers==6.0.0a

purl==0.8
pydelicious==0.6
pymongo==2.6.2
pyparsing==2.0.2
python-dateutil==2.2
pytz==2014.4
redis==2.10.1
requests==1.2.0

sentry-sms-ru==0.0.3
six==1.7.2
sorl-thumbnail==11.12
sqlparse==0.1.11
stevedore==0.15
surlex==0.2.0
urllib3==1.8.2




wsgiref==0.1.2
